Create node project with typescript and jest
The package.json file was created with placeholder metadata (version, license, ...) and some useful scripts to get started with development (start, build, test).

The following devDependencies were added:

- Node.js basics for typescript projects: typescript, ts-node, @types/node

- Testing for typescript projects: jest, ts-jest, @types/jest

The tsconfig.json and jest.config.js files configure the typescript building and jest testing tasks and are setup in such a way that the source code from the src/ directory is compiled into the dist/ directory. The test files with name pattern *.spec.ts are excluded from compilation, but are picked up by the test runner. It is assumed that for each *.ts file there is a accompanying *.spec.ts file.

The src/main.ts and src/main.spec.ts files are examples for such files and contain placeholder code to verify that the npm scripts (start, build, test) are working correctly.

The .gitignore contains the dist/, node_modules/ and coverage/ directories that contain generated code or dependencies and should never be committed. Also, it contains the .idea/ folder, which contains the settings of the IDE that this project was created with (Webstorm).