Add production builds, license, readme and metadata in package.json
The build process: for building release versions, a build:prod run script was added. This script runs the regular build command, but switches out the tsconfig.json file for a special tsconfig.prod.json file that emits the transpilation result in a ready-to-release format. The rimraf dependency was added to be able to implement an operating system independent clean script (using rimraf instead of rm -rf, which would only work on linux based systems).

Linting: a set of rules was added to enforce a specific jsdoc schema to ensure that all code is not only well-formatted, but also fully documented. To achieve this, the eslint-plugin-jsdoc was added and configured using the .eslintrc.js

A prepublish script was added to run the full set of tools before attempting to publish the library.

The placeholder metadata in the package.json was replaces with real content meant for publication. This includes for example the author and license. The "main" and "types" entries now point to the index file of this library.

The main.ts file and its tests were completely removed in favor if the index.ts file, which is required for building libraries. It is referenced in the package.json file as mentioned above.