Migrate to npm workspace
Create a npm workspace and move all existing code into the packages/color-identity package. Documentation and scripts have been updated accordingly to keep working, however a long term strategy for how to maintain releases of multiple packages in this workspace must be developed. For now, the existing package can be released as usual and the workspace is ready for new packages to be prototyped.